# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OmniRoute API
  version: 3.8.50
  description: |
    OmniRoute is a local-first AI API proxy router. It provides an OpenAI-compatible
    endpoint that routes requests to multiple AI providers with load balancing,
    failover, and usage tracking.

    ## Base URLs
    - **Local**: `http://localhost:20128`

    ## Authentication
    All proxy endpoints require a Bearer token (API key managed via the dashboard).
    Management endpoints are protected when `requireLogin` is enabled.
  contact:
    name: OmniRoute
  license:
    name: MIT
servers:
  - url: http://localhost:20128
    description: Local development
...
paths:
...
  /api/combos:
    get:
      tags: [Combos]
      summary: List routing combos
      responses:
        "200":
          description: Combo list
    post:
      tags: [Combos]
      summary: Create routing combo
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ComboCreate"
      responses:
        "201":
          description: Created combo
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: API key obtained from the OmniRoute dashboard
    ManagementSessionAuth:
      type: apiKey
      in: cookie
      name: auth_token
      description: Dashboard management session cookie for protected management routes
...
  schemas:
...
    ComboCreate:
      type: object
      required: [name]
      properties:
        name:
          type: string
          maxLength: 100
        description:
          type: string
          maxLength: 2000
        strategy:
          type: string
          enum:
            - priority
            - weighted
            - round-robin
            - context-relay
            - fill-first
            - p2c
            - random
            - least-used
            - cost-optimized
            - reset-aware
            - reset-window
            - headroom
            - strict-random
            - auto
            - lkgp
            - context-optimized
            - cache-optimized
            - fusion
            - pipeline
          default: priority
        models:
          type: array
          items:
            oneOf:
              - type: string
              - $ref: "#/components/schemas/ComboModelStep"
              - $ref: "#/components/schemas/ComboRefStep"
        config:
          $ref: "#/components/schemas/ComboRuntimeConfig"
...
    ComboModelStep:
      type: object
      required: [model]
      properties:
        id:
          type: string
          description: Stable Step ID used by adaptive role references.
        kind:
          type: string
          enum: [model]
        provider:
          type: string
        providerId:
          type: string
        model:
          type: string
        connectionId:
          type: string
          nullable: true
        weight:
          type: number
          minimum: 0
          maximum: 100
    ComboRefStep:
      type: object
      required: [kind, comboName]
      properties:
        id:
          type: string
        kind:
          type: string
          enum: [combo-ref]
        comboName:
          type: string
        weight:
          type: number
          minimum: 0
          maximum: 100
    ComboScoringWeights:
      type: object
      additionalProperties: false
      required: [quota, health, costInv, latencyInv, taskFit, stability]
      properties:
        quota: { type: number, minimum: 0, maximum: 1 }
        health: { type: number, minimum: 0, maximum: 1 }
        costInv: { type: number, minimum: 0, maximum: 1 }
        latencyInv: { type: number, minimum: 0, maximum: 1 }
        taskFit: { type: number, minimum: 0, maximum: 1 }
        stability: { type: number, minimum: 0, maximum: 1 }
        tierPriority: { type: number, minimum: 0, maximum: 1 }
        tierAffinity: { type: number, minimum: 0, maximum: 1 }
        specificityMatch: { type: number, minimum: 0, maximum: 1 }
        contextAffinity: { type: number, minimum: 0, maximum: 1 }
        cacheAffinity: { type: number, minimum: 0, maximum: 1 }
        sessionAvailability: { type: number, minimum: 0, maximum: 1 }
        resetWindowAffinity: { type: number, minimum: 0, maximum: 1 }
    ComboRuntimeConfig:
      type: object
      additionalProperties: true
      properties:
        candidatePool:
          type: array
          items:
            type: string
        fastWorkerModelRefs:
          type: array
          description: Stable model Step IDs preferred for simple requests.
          maxItems: 200
          items:
            type: string
        strongReasoningModelRefs:
          type: array
          description: Stable model Step IDs preferred for complex requests.
          maxItems: 200
          items:
            type: string
        adaptiveJudgeModelRef:
          type: string
          description: Optional single model Step ID used to classify each request before adaptive role selection.
          minLength: 1
          maxLength: 200
        weights:
          description: Default scoring profile for neutral requests and roles without a custom profile.
          $ref: "#/components/schemas/ComboScoringWeights"
        fastWorkerWeights:
          description: Optional scoring profile used to rank multiple eligible Fast Worker Steps.
          $ref: "#/components/schemas/ComboScoringWeights"
        strongReasoningWeights:
          description: Optional scoring profile used to rank multiple eligible Strong Reasoning Steps.
          $ref: "#/components/schemas/ComboScoringWeights"
        modePack:
          type: string
        routerStrategy:
          type: string
        explorationRate:
          type: number
          minimum: 0
          maximum: 1
        budgetCap:
          type: number
          exclusiveMinimum: 0