# src/app/api/combos/route.ts
import { NextResponse } from "next/server";
import { createCombo, getComboByName, getCombos, getCombosCount } from "@/lib/db/combos";
import { isCloudEnabled } from "@/lib/db/settings";
import { getConsistentMachineId } from "@/shared/utils/machineId";
import { syncToCloud } from "@/lib/cloudSync";
import { validateCompositeTiersConfig } from "@/lib/combos/compositeTiers";
import { normalizeComboModels } from "@/lib/combos/steps";
import { validateComboDAG, clampComboDepth } from "@omniroute/open-sse/services/combo.ts";
import { createComboSchema, paginationSchema } from "@/shared/validation/schemas";
import { isValidationFailure, validateBody } from "@/shared/validation/helpers";
import { requireManagementAuth } from "@/lib/api/requireManagementAuth";
import { comboErrorResponse } from "@/lib/api/comboErrorResponse";
import { computeComboContextLength } from "@/lib/combos/comboContext";
import { normalizeIntelligentRolePoolConfig } from "@/lib/combos/intelligentRouting";
import { ComboInvariantError } from "@/lib/combos/invariants";
import { buildComboNameCollisionWarning } from "@/lib/combos/modelNameCollision";

// GET /api/combos - Get all combos
export async function GET(request: Request) {
  const authError = await requireManagementAuth(request);
  if (authError) return authError;

  try {
    const { searchParams } = new URL(request.url);
    const raw = {
      offset: searchParams.get("offset") || undefined,
      limit: searchParams.get("limit") || undefined,
    };
    const validation = validateBody(paginationSchema, raw);
    if (isValidationFailure(validation)) {
      return NextResponse.json({ error: validation.error }, { status: 400 });
    }

    const range = validation.data;
    const total = getCombosCount();
    const rawCombos = await getCombos(range.limit, range.offset);
    const combos = rawCombos.map((combo) => ({
      ...combo,
      computed_context_length: computeComboContextLength(combo, rawCombos),
    }));
    return NextResponse.json({ combos, total });
  } catch (error) {
    console.log("Error fetching combos:", error);
    return NextResponse.json({ error: "Failed to fetch combos" }, { status: 500 });
  }
}

// POST /api/combos - Create new combo
export async function POST(request) {
  const authError = await requireManagementAuth(request);
  if (authError) return authError;

  try {
    const body = await request.json();

    // Zod validation (covers name format, length, etc.)
    const validation = validateBody(createComboSchema, body);
    if (isValidationFailure(validation)) {
      return NextResponse.json({ error: validation.error }, { status: 400 });
    }
    const allCombos = await getCombos();
    const normalizedModels = normalizeComboModels(validation.data.models, {
      comboName: validation.data.name,
      // `allCombos` from `getCombos()` is typed as the DB-shaped record
      // (JsonRecord & { version: 2; models: ComboStep[] }) which is
      // structurally compatible with the local ComboCollectionLike in
      // `normalizeComboModels` but TS does not infer the relationship.
      allCombos: allCombos as never,
    });
    const comboInput = {
      ...validation.data,
      models: normalizedModels,
      ...(validation.data.config
        ? {
            config: normalizeIntelligentRolePoolConfig(validation.data.config, normalizedModels),
          }
        : {}),
    };
    const { name, strategy, config } = comboInput;
    const compositeValidation = validateCompositeTiersConfig(comboInput);
    if (compositeValidation.success === false) {
      const failure = compositeValidation as {
        success: false;
        error: { message: string; details: unknown[] };
      };
      return comboErrorResponse(
        "COMBO_003",
        400,
        { reason: failure.error.message, details: failure.error.details },
        request
      );
    }

    // Check if name already exists
    const existing = await getComboByName(name);
    if (existing) {
      return NextResponse.json({ error: "Combo name already exists" }, { status: 400 });
    }

    // Validate nested combo DAG (no circular references, max depth)
    // Temporarily add the new combo to validate its graph
    const tempCombo = {
      ...comboInput,
      name,
      strategy,
      config,
    };
    try {
      validateComboDAG(
        name,
        [...allCombos, tempCombo],
        new Set(),
        0,
        clampComboDepth((config as { maxComboDepth?: unknown } | undefined)?.maxComboDepth)
      );
    } catch (dagError) {
      return NextResponse.json({ error: dagError.message }, { status: 400 });
    }

    const combo = await createCombo(comboInput);

    // Auto sync to Cloud if enabled
    await syncToCloudIfEnabled();

    // #8530: a combo named after a real model id is a supported pattern
    // (#6940 — bare-model-id provider fallback), so it is never rejected.
    // Surface it as a non-blocking warning so the dashboard/API caller can
    // confirm it was intentional instead of silently shadowing the model.
    const warning = buildComboNameCollisionWarning(name);
    return NextResponse.json(warning ? { ...combo, warning } : combo, { status: 201 });
  } catch (error) {
    if (error instanceof ComboInvariantError) {
      return comboErrorResponse("COMBO_008", 400, { reason: error.message }, request);
    }
    console.log("Error creating combo:", error);
    return NextResponse.json({ error: "Failed to create combo" }, { status: 500 });
  }
}

/**
 * Sync to Cloud if enabled
 */
async function syncToCloudIfEnabled() {
  try {
    const cloudEnabled = await isCloudEnabled();
    if (!cloudEnabled) return;

    const machineId = await getConsistentMachineId();
    await syncToCloud(machineId);
  } catch (error) {
    console.log("Error syncing to cloud:", error);
  }
}